--- a/tests/testthat/test-sample.R
+++ b/tests/testthat/test-sample.R
@@ -98,20 +98,6 @@ test_that("can't sample more values than obs (without replacement)", {
   )
 })
 
-test_that("sampling grouped tbl samples each group with different replacement policies", {
-  by_cyl <- mtcars %>% group_by(cyl)
-
-  group_sizes <- group_size(by_cyl)
-  sample_size <- median(group_sizes)
-  with_replacements <- group_sizes < sample_size
-
-  sampled <- by_cyl %>% sample_n(sample_size, with_replacements)
-  expect_is(sampled, "grouped_df")
-  expect_groups(sampled, "cyl")
-  expect_equal(nrow(sampled), sample_size * 3)
-  expect_equal(sampled$cyl, rep(c(4, 6, 8), each = sample_size))
-})
-
 
 df2 <- data.frame(
   x = rep(1:2, 2),