@saurfang 2017-11-06
Sure. I do think there is a legitimate use case for different sample sizes per group but I agree that the interface does not feel natural here given the vector relies on an internal knowledge of the group order. This PR is a little too messy and I started a fresh one at #3193.

I am happy to submit followup PRs if you think it is worth tidying up the R/C++ index difference and possible corner cases when `indices` are not available.